Allow for default variable interpolation notation
Previously, to use variable interpolation with this consul.hcl.j2
template and the other templates located in the `templates/ctpl` folder,
the used variables would have to be referenced with '[[' and ']]'
instead of the standard jinja2 template '{{' and '}}' notation. This was
due to the fact that the jinja2 templates referenced are Go templates
which use the same '{{' and '}}' variable notation.

This commit adds `set_fact` tasks that create '_'-prefixed variable
equivalents of the variables used in the Go templates of this role. This
safely deals with the variable interpolation conundrum and allows for
users to use standard '{{' and '}}' variable interpolation for their own
variables.