fix(config): add chosen interface to client_addr
This change will expose Consul to the local network on the host's main
interface. While this is a potential security risk, it was needed as
Prometheus is not able to collect metrics from Consul agents without
this configuration.

This will likely only be a temporary change and will be changed once I
have thought of a better way to expose Consul metrics to Prometheus.